change update module command name to upgrade module
