+++ src/Core/Domain/Module/Exception/CannotUpgradeModuleException.php
@@ -24,33 +24,8 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-declare(strict_types=1);
+namespace PrestaShop\PrestaShop\Core\Domain\Module\Exception;
 
-namespace PrestaShop\PrestaShop\Core\Domain\Module\Command;
-
-use PrestaShop\PrestaShop\Core\Domain\Module\ValueObject\ModuleTechnicalName;
-
-/**
- * Update module
- */
-class UpdateModuleCommand
+class CannotUpgradeModuleException extends ModuleException
 {
-    private ModuleTechnicalName $technicalName;
-
-    /**
-     * @param string $technicalName Technical name for module
-     */
-    public function __construct(string $technicalName)
-    {
-        $this->technicalName = new ModuleTechnicalName($technicalName);
-    }
-
-    /**
-     * @return ModuleTechnicalName
-     */
-    public function getTechnicalName(): ModuleTechnicalName
-    {
-        return $this->technicalName;
-    }
-
 }
