+++ src/PrestaShopBundle/Resources/config/services/adapter/module.yml
@@ -26,6 +26,11 @@ services:
     autowire: true
     public: false
 
+  PrestaShop\PrestaShop\Adapter\Module\CommandHandler\UpgradeModuleHandler:
+    autoconfigure: true
+    autowire: true
+    public: false
+
   PrestaShop\PrestaShop\Adapter\Module\CommandHandler\BulkUninstallModuleHandler:
     autoconfigure: true
     autowire: true
