+++ tests/Integration/Behaviour/Features/Scenario/Module/upgrade_module.feature
@@ -1,15 +1,15 @@
-# ./vendor/bin/behat -c tests/Integration/Behaviour/behat.yml -s module --tags update-module
+# ./vendor/bin/behat -c tests/Integration/Behaviour/behat.yml -s module --tags upgrade-module
 @reset-test-modules-before-feature
 @restore-all-tables-before-feature
 @clear-cache-before-feature
 @reset-test-modules-after-feature
-@update-module
+@upgrade-module
 Feature: Module
   PrestaShop allows BO users to manage modules
   As a BO user
-  I must be able to upload/update modules
+  I must be able to upload/upgrade modules
 
-  Scenario: Update module with zip file from remote (not present initially)
+  Scenario: Upgrade module with zip file from remote (not present initially)
     # Check that module is not present before we upload it
     Given module dashproducts has following infos:
       | installed | false |
@@ -35,7 +35,7 @@ Feature: Module
       | module_version    | 2.1.4        |
       | enabled           | true        |
       | installed         | true         |
-    And I update module dashproducts
+    And I upgrade module dashproducts
     Then module dashproducts has following infos:
       | technical_name    | dashproducts |
       | installed_version | 2.1.4        |
@@ -44,7 +44,7 @@ Feature: Module
       | installed         | true         |
 
 
-  Scenario: Update module with zip file from remote (present initially)
+  Scenario: Upgrade module with zip file from remote (present initially)
     # Check that module is not present before we upload it
     Given module dashproducts has following infos:
       | technical_name    | dashproducts |
@@ -59,7 +59,7 @@ Feature: Module
       | module_version    | 2.1.3        |
       | enabled           | false        |
       | installed         | false        |
-    And I update module dashproducts
+    And I upgrade module dashproducts
     Then module dashproducts has following infos:
       | technical_name    | dashproducts |
       | installed_version | 2.1.4        |
